Add code to persist some dummy posts and comments
Notice the use of @PostConstruct in main class. This is deliberately
done to avoid writing rest controllers to test the application.

Also, cascade type and fetch type is added to post and comments
relationship.

CascadeType.ALL means any operation to parent(post) should also
be cascaded to the child. Eg. if we save a post with some comments
set, it would save the comments also.

FetchType.EAGER means to fetch comments also when we fetch post.